+++ CHANGELOG.md
@@ -33,6 +33,9 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 * Refactor args in `create_optimization_grid` to be consistent with
   `optimize(...)`.
 * Fix response curves for KPI-based optimization.
+* Add `plot_channel_contribution_area_chart` method to `MediaSummary` in the
+  visualizer.
+* Add organic media support for adstock decay in analyzer.
 
 ## [1.0.7] - 2025-03-19